fix(vercel): remove win32 rollup dep and pin engines.node=20.x; refactor CSRF+rate limiting to shared /security (dev-only limiter); fix archived doc typo
