+++ PROJECT_STATUS_COMPLETE.md
@@ -1,3 +1,3 @@
-y6vs document has been archived.
+ This document has been archived.
 
 See: `docs/archive/PROJECT_STATUS_COMPLETE.md`
