+++ src/routes/api/merge-chunks/+server.ts
@@ -1,6 +1,7 @@
 import { json, type RequestHandler } from '@sveltejs/kit';
 import type { TranscriptionResult } from '../../../contracts/transcription';
 import { ConsensusComparisonEngine } from '../../../implementations/comparison';
+import { checkRateLimit, getClientIP, validateCsrfFromJson } from '$lib/security';
 
 /*
   Merge per-chunk per-service transcripts into a final consensus-like result.
@@ -12,93 +13,11 @@ import { ConsensusComparisonEngine } from '../../../implementations/comparison';
   Strategy:
   - For each service, concatenate its chunk texts in order (with spaces).
   - Build synthetic TranscriptionResult[] with one result per service.
-// --- simple dev-only rate limit (same pattern as /api/transcribe) ---
-const RATE_LIMIT_CONFIG = {
-  MAX_REQUESTS: 20,
-  WINDOW_MS: 15 * 60 * 1000 // 15 minutes
-};
-const rateLimitMap = new Map<string, { count: number; windowStart: number }>();
-
-function getClientIP(request: Request): string | null {
-  const xff = request.headers.get('x-forwarded-for');
-  if (xff) return xff.split(',')[0].trim();
-  const xri = request.headers.get('x-real-ip');
-  if (xri) return xri.trim();
-  return null;
-}
-
-function cleanupExpiredEntries(now: number) {
-  for (const [ip, data] of rateLimitMap.entries()) {
-    if (now - data.windowStart >= RATE_LIMIT_CONFIG.WINDOW_MS) rateLimitMap.delete(ip);
-  }
-}
-
-function checkRateLimit(clientIP: string | null): { allowed: boolean; retryAfter?: number; error?: string } {
-  if (!clientIP) return { allowed: false, error: 'Could not determine client IP address for rate limiting' };
-  const now = Date.now();
-  cleanupExpiredEntries(now);
-  const entry = rateLimitMap.get(clientIP);
-  if (!entry) {
-    rateLimitMap.set(clientIP, { count: 1, windowStart: now });
-    return { allowed: true };
-  }
-  if (now - entry.windowStart >= RATE_LIMIT_CONFIG.WINDOW_MS) {
-    rateLimitMap.set(clientIP, { count: 1, windowStart: now });
-    return { allowed: true };
-  }
-  if (entry.count < RATE_LIMIT_CONFIG.MAX_REQUESTS) {
-    entry.count++;
-    return { allowed: true };
-  }
-  const timeRemaining = RATE_LIMIT_CONFIG.WINDOW_MS - (now - entry.windowStart);
-  return { allowed: false, retryAfter: Math.ceil(timeRemaining / 1000) };
-}
 
   - Run the existing comparison engine to produce a ConsensusResult.
   Caveat: token-level confidences won’t exist here; this is intended for large-file assembly.
 */
 
-// --- simple dev-only rate limit (same pattern as /api/transcribe) ---
-const RATE_LIMIT_CONFIG = {
-  MAX_REQUESTS: 20,
-  WINDOW_MS: 15 * 60 * 1000 // 15 minutes
-};
-const rateLimitMap = new Map<string, { count: number; windowStart: number }>();
-
-function getClientIP(request: Request): string | null {
-  const xff = request.headers.get('x-forwarded-for');
-  if (xff) return xff.split(',')[0].trim();
-  const xri = request.headers.get('x-real-ip');
-  if (xri) return xri.trim();
-  return null;
-}
-
-function cleanupExpiredEntries(now: number) {
-  for (const [ip, data] of rateLimitMap.entries()) {
-    if (now - data.windowStart >= RATE_LIMIT_CONFIG.WINDOW_MS) rateLimitMap.delete(ip);
-  }
-}
-
-function checkRateLimit(clientIP: string | null): { allowed: boolean; retryAfter?: number; error?: string } {
-  if (!clientIP) return { allowed: false, error: 'Could not determine client IP address for rate limiting' };
-  const now = Date.now();
-  cleanupExpiredEntries(now);
-  const entry = rateLimitMap.get(clientIP);
-  if (!entry) {
-    rateLimitMap.set(clientIP, { count: 1, windowStart: now });
-    return { allowed: true };
-  }
-  if (now - entry.windowStart >= RATE_LIMIT_CONFIG.WINDOW_MS) {
-    rateLimitMap.set(clientIP, { count: 1, windowStart: now });
-    return { allowed: true };
-  }
-  if (entry.count < RATE_LIMIT_CONFIG.MAX_REQUESTS) {
-    entry.count++;
-    return { allowed: true };
-  }
-  const timeRemaining = RATE_LIMIT_CONFIG.WINDOW_MS - (now - entry.windowStart);
-  return { allowed: false, retryAfter: Math.ceil(timeRemaining / 1000) };
-}
 
 export const POST: RequestHandler = async ({ request, cookies }) => {
   try {
@@ -118,12 +37,12 @@ export const POST: RequestHandler = async ({ request, cookies }) => {
     }
 
     // CSRF: double-submit token validation (body token must match cookie)
-    const csrfToken = (body?.csrfToken ?? '') as string;
     const csrfCookie = cookies.get('csrfToken');
-    if (!csrfToken || !csrfCookie || csrfToken !== csrfCookie) {
+    if (!validateCsrfFromJson(body, csrfCookie)) {
       return json({ error: 'Request could not be processed. Please try again.' }, { status: 403 });
     }
-    const chunkTexts = Array.isArray(body?.chunkTexts) ? body.chunkTexts : [];
+    const chunkTexts: Array<{ index: number; textsByService: Record<string, string> }>
+      = Array.isArray(body?.chunkTexts) ? body.chunkTexts : [];
     if (chunkTexts.length === 0) {
       return json({ error: 'No chunk texts provided' }, { status: 400 });
     }
